Make it cheaper to access a `ValueReference`.
Since we know that if a `ValueReference` holds a raw pointer it must
always be valid we can inline some paths from `_get` directly into
`_safeGet`. We also reduce the size of `_get` from the knowledge that we
must hold exactly two alternatives, and replace uses of `variant`
`std::get` (which sets up exception handling) with `std::get_if` which
is slightly cheaper at runtime.