Add new `skip` unit item that efficiently skips over uninteresting data.
`skip` supports the attributes `&size`, `&eod`, and `&until` with the
usual semantics. `skip` takes the role of what used to be parsing
fields of type `void`, which we will deprecate in a subsequent commit.

We implement `skip` through a new unit item class, along with a
corresponding production. Doing so allows us to keep its parsing out
of the normal (pretty complex) field processing, and instead generate
straight code just for this case. This also allows to simply the field
code a bit.

Note that we do not make "skip" a new keyword in the parser to avoid
clashing with usage in existing scripts (incl. in the test suite). So
this stays backwards compatible.