Make sure Spicy runtime config is initialized after `spicy::rt::init`.
We previously wouldn't explicitly initialize the Spicy runtime
configuration but instead rely on `configuration::get` (which implicitly
initializes it) to be called somewhere in the driver. This makes it hard
to use parsers outside of our drivers, we e.g., saw nullptr derefs in
our fuzzer binaries.

With this patch we force initialization in `spicy::rt::init` which is
the documented way to ensure all runtime dependencies are set up
correctly.