Remove `void` parsing with any attributes.
We no longer support `void` parsing with `&eof`, `&size`, or `&until`
because `skip` has taken that role. The only use of `void` that's
still supported is without any attributes at all, which is essentially
no-op (useful in particular with `switch`). This change allows us to
remove a couple of special-cases for `void` during parser generation
as well.

For now, this changes breaks our existing tests that are using those
attributes with void. In a subsequent commit we will actually bring
back temporary, deprecated support for void-with-attributes by mapping
it to `skip`, and then the tests will work again.