Allow `skip` for all fields.
Mulling further over the new syntax, I think it's confusing to support
`skip` for some types of fields but not others; seems hard for users
to understand what works and what doesn't. This commit makes that
difference an implementation detail: we create specialized code for
fields where we support it, and just fall back to the standard parsing
where not.