Add support for `skip` to literals.
All literals now get a separate `skip` path. It's a bit unclear how
much this buys us in the end, but it was easy to do, fits in well, and
might be helpful in particular for regular expressions.

The implementation leverages the existing literal parsing code, but
introduce a new `Skip` mode that doesn't save any values. Note that
unfortunately we still need to include the look-ahead logic. Would be
nicer if we had a way to tell statically whether look-ahead parsing is
active (but that's independent of `skip`).