Fix readthedocs build.
We previously were not using a readthedocs config file which causes the
build to fail due to an unpinned urllib3 (good) or an outdated system
(bad), see https://github.com/readthedocs/readthedocs.org/issues/10290.

With this patch we use the recommended approach which a config file
which gives us control over the build environment. We also bump docs
dependencies in order to fix version incompatibilities triggered by the
new setup, see
https://github.com/readthedocs/readthedocs.org/issues/9876.