Do not forcibly make `strong_ref` `in` function parameters immutable.
We previously would always emit a `const` value when passing a
`strong_ref` as a function parameter. This lead to the unintuitive error
that non-const `strong_ref` parameters could not be modified or have
non-const methods called on them (this would actually fail at codegen
time).

With this patch we remove the special treatment of `strong_ref`
parameter mutability.

Closes #1447.