Update Zeek docs for Zeek 6.0.
With the plugin code now being part of Zeek proper, the main change is
getting rid of all references to a "plugin". However, I've left the
text referring to Zeek 5.0 as the one that started shipping Spicy
built-in, since from a user's perspective there's not much difference
between 5.0 and 6.0 in that regard.