Reject uses of `self` in unit `&size` and `&max-size` attribute.
Values in `self` are only available after parsing has started while
`&size` and `&max-size` are consumed before that. This means that any
use of `self` and its members in these contexts would only ever see
unset members, so it should not be the intended use.

This patch adds a validation which rejects uses of `self` for these unit
attributes.

Closes #1466.