Implement remaining pieces to make `bitfield` a full HILTI type.
We keep representing bitfields as a C++ tuple behind the scenes, but
no longer map it into a HILTI-side tuple as well. Instead `Bitfield`
is now a regular type with corresponding operators and runtime
functionality.

The one change to the C++ tuple representation is that we add an
additional hidden tuple element at the end that stores the original
integer value. That allows for displaying it (which we will do for now
only inside Spicy debug output) and could later also facilitate modifying
fields and reconstructing an integer.