Use user-inaccessible chars for encoding `::` in feature variables.
When setting up feature tracking variables for the optimizer we
previously would normalize `:` as `_`, e.g., `mod::Unit` would lead to
feature variables `__feat%mod__Unit%...`. In various places in the
optimizer we would then use that knowledge to extract module and unit
ID. Since user-specified identifiers can contain literal `_` or `__`
this could confuse the optimizer.

With this patch we now normalize `:` as `@` which is not
user-accessible. We add a new codegen normalization so `@` becomes
`0x40`. This gives a more reliable encoding.

Closes #1504.