Fix when GH `email-failure` action sends notification.
In order to detect whether to send out a failure email we previously
would check that the action conclusion contained certain statuses. This
check seems to have been nonsensical as we only notified for benign
failure reasons (the check would have made a little more sense if
negated, but would have then also notified on successes).

With this patch we instead explicitly enumerate the failure conclusions
to trigger on, see
https://docs.github.com/en/webhooks/webhook-events-and-payloads#check_suite--check_suite-object
for the full list of conclusions.