Handle anonymous bitfields inside `switch` statements.
We now map items of anonymous bitfields inside a `switch` cases into
the unit namespace, just like we already do for top-level fields. We
also catch if two anonymous bitfields inside those cases carry the
same name, which would make accesses ambiguous.

So the following works now:

```
    switch (self.n) {
        0 -> : bitfield(8) {
            A: 0..7;
        };
        * -> : bitfield(8) {
            B: 0..7;
        };
    };
```

Whereas this does not work:

```
    switch (self.n) {
        0 -> : bitfield(8) {
            A: 0..7;
        };
        * -> : bitfield(8) {
            A: 0..7;
        };
    };
```

The latter not working is reasonable I think, rather than trying to
figure out that the two cases have the same type and hence could be
mapped to a single data member in the resulting code (which we do, and
support, for cases of the same *name* and type).

Closes #1503.