Add begin to parser state.
This patch adds the current begin position to the parser state, and
makes the corresponding changes to generated parser functions so it is
passed down.

We already modelled the semantic beginning of the input in the unit, but
had no reliable way to keep this up-to-date across non-unit contexts
like `&parse-from`. This would then for certain setups lead to generated
code where `input` and `position` would point to different inputs which in
turn caused `offset` (modelled as `position - input`) to be incorrect.

(cherry picked from commit 35aa4266c8bf37ebff3df5ec17241171e42e7cec)