Declare Spicy pygments extension as parallel-safe. [skip CI]
We previously would not declare that the Spicy pygments highlighter is
safe to execute in parallel (reading or writing of sources). Sphinx then
assumed that the extension was not safe to run in parallel and instead
ran jobs sequentially.

This patch declares the extension as able to execute in parallel. Since
the extension does not manage any external state this is safe.

(cherry picked from commit 22e9f729f423d0e43b788d48d824377ceaaf8ede)