Add explicit dependency for local pre-commit hook.
The implementation of the `stray-baselines` hook relies on `btest`. We
would previously assume it was installed out of band which makes the
default pre-commit hooks cumbersome to use (e.g., opaque failure if
dependency missing).

This patch explicitly declares the Python dependency which then gets
managed by pre-commit. This should come with no new risk as we do not
rely on version-dependent BTest features in the hook.