Fix potential infinite loop when trimming data before stream.
Previously we would trigger an infinite loop if one tried to trim before
the head chunk of a stream. In praxis this seem to have been no issue
due to #1549 and us emitting way less calls to trim than possible.

This patch adds an explicit check whether we need to trim anything, and
exits the low-level function early for such cases.

Closes #1554.