Explicitly set ASM compiler.
When not setting an ASM compiler we might end up configuring Clang as C
compiler, but GCC as ASM compiler. The code in `3rdparty/fiber` then
causes Clang-only flags like `-Weverything` to be passed to GCC (which
does not understand it).

The way we need to do this is subtle. We need to rely on
`3rdparty/fiber` to explicitly `enable_language(ASM)` and not set it as
a project language to work around `find_package(BISON)` breaking if
`CMAKE_ASM_COMPILER` is set. By removing ASM from project languages and
setting the ASM config after Bison was found, but before we configure
`3rdparty/fiber` things seem to work as intended.