Use shared_ptr with non-atomic refcounts in the runtime library.
We heavily use `std::shared_ptr` in the runtime library in types which
we expect to be cheap to copy, e.g., in views or safe iterators. By
design `std::shared_ptr` has to use atomic ints for its strong and weak
counts. Since parsing is single threaded the use of atomics can
introduce overhead without any need.

This patch replaces uses of `std::shared_ptr` in the runtime library
with a shared_ptr without atomic refcounts. Since we still need weak
pointers this has to be a slightly more complete replacement.