Back Chunk with a std::string instead of std::variant.
This is probably a bit funky due to the trailing \0 and a Chunk not
being a string, but hey, hilti::rt::Bytes is backed by std::string, too.

Looking at the creation of Chunk::Chunk(), there's actually a memset()
operation visible due to the variant holding some 40 bytes that are
zero initialized upon construction. std::string supports SSO, so we
can leverage even if the actual size is out of our control. The
std::get_if()s do show up very marginally for the View::size() calls.