Display floats in scientific notation in profiler report.
We previously would display floats as decimals with two decimals in the
profiler report. While this makes it easy to get a quick overview of the
magnitude from directly looking at the logs, in real parsers many more
components might be involved which each only contribute less than 1%;
such contributions are currently displayed as 0.00.

This patch switches display of floating point numbers in the profiler
report to use scientific notation so we do not loose information for
small fractions.