Optimize C++ tuple element coercions.
This patch introduces an optimization to pass C++ tuples through if they
do not need C++-side coercion. This emits better code if a tuple ctor
already has the right types. We also optimize codegen if we are coercing
from a temporary tuple in that we now capture it in a temporary and
coerce elements from that instead of repeatedly emitting the same tuple
ctor (we scale with the number of tuple elements now not its square
anymore).

Closes #1615.