Fix handling of `%synchronize-*` attributes for units in lists.
We previously would not detect `%synchronize-at` or `%synchronize-from`
attributes if the unit was not directly in a field, i.e., we mishandled
the common case of synchronizing on a unit in a list.

With this patch we now handle these attributes, regardless of how the
unit appears.

Closes #1617.

(cherry picked from commit 57debb2d3abb3a89690d10991b1ddfd5e3198e14)