Add move ctr for `stream::View`.
In C++17 move ctors are not by default generated, add them explicitly.
We also get rid of the `View` dtr since it adds nothing (e.g., the class
is declared `final`), but declaring them could e.g., make the class not
POD anymore (this is not the case here though).