Remove use of `IntrusivePtr` over `const`.
It looks like using a `IntrusivePtr<const T>` interferes with default'ed
move constructors and assignment operators so that the non-move versions
are called. This can incur additional overhead.

This patch simply gets rid all `IntrusivePtr<const T>` in favor of
`IntrusivePtr<T>` which does not seem to show this issue. It might be
possible to instead adjust something e.g., `ManagedObject` so
`IntrusivePtr` does not regress when holding semantically `const`
values though.

For a big internal parser this shows runtime improvements up to 2%.