Fix Bison locations.
If a Bison rule started with an optional element, its location would
start at the end of the *previous* token if that optional element
wasn't present. We now skip token locations that are zero-sized at the
beginning of a rule. This isn't pretty, but the best I could come up
with. I spot-checked the locations triggering baseline differences in
our tests, and all I looked at are strictly better than before.