Fix handling of `spicyc -x` when passed a filename.
For the `-x` flag users can pass either the name of a directory or a
file. If passed a filename the file name is incorporated into the
identifier generated namespace identifier of e.g., the form
`__hlt_FILENAME` in the emitted C++ files. Previously when passing a
directory we would still emit a `_` separator, but since there was no
filename nothing after it so we generated namespace identifiers like
`__hlt_`.

With this patch we now only emit the separator if passed a filename.