Speed up `util::toIdentifier`.
Turns out this is called a lot. This changes baselines due to slight
differences in generated IDs (changes are good because they better avoid
pontential conflicts).

This also removes the option to ensure non-keyword IDs: I don't
think we need that because we check for that when we generate C++
code.