Fix constness validation for unit field assignments.
The existing validator was to eager to reject legitimate code. The fix
is two parts: (1) remove that validator, and (2) ensure that the
const-version of unit field assignments produces a result with the
correct const/rhs semantics. The latter we can then check more
directly with an improved validator for the `assign` statement. With
that, it turns out that we don't need some special-casing logic
anymore that previously manually replaced field accesses with an
operator.