Remove some dead statements.
This patch adds a component to compute control and data flow for
statements. We then use this component to detect "dead statements",
i.e., statements which are not seen by any other needed computations.
Currently we handle two classes of dead statements:

- assignments which are override before being used
- unreachable code, e.g., due to preceding `return` or `throw`

While the latter is pretty straightforward, the former becomes tricky
due to exceptions. In Spicy pretty much any expression could throw, and
without fully modelling this control flow becomes practically
unmanageable. We work around this by ignore possibly throwing
expressions when constructing the control flow graph, but instead taking
into account whether the effect of a statement is actually visible. In
that framework to decide whether a write to a block-local declaration is
visible we can use regular data flow; for "globally visible data" on the
other hand we always need to consider each write potentially visible. We
need to treat globals, `inout` parameters and struct members as globally
visible.