Fix fuzzer builds for reworked AST.
When generating fuzzer binaries we previously would generate parser and
linker code in two invocations, one for the module's C++ code with `-c`
and one for the linker code with `-l` As of
28437292fcf01cb1602e0318924cd7c78f5429fe this generates code which
often fails to link due to missing symbols (this seems to be due to us
not including code from the `filter` module in the single-file C++
output anymore).

This patch reworks to code to switch from using two separate `spicyc`
invocations with `-c` and `-l` to emitting all sources at once with
`-x`. This requires some changes to the CMake function generating the
targets as we need to statically name all inputs to `add_executable`
which generates code for each module and the linker file. We also rename
the binaries in the process.