Rework order of C++ codegen.
We used to potentially codegen modules multiple times. Now we cache a
module's generated C++ code inside the AST node the first time we
create it. From there, we can then easily reuse it later, in
particular when needing to import its declarations into another
module. Internally, we switch storage for `cxx::Unit` to shared
pointers, and tweak the debug logging a bit for better readability in
this new model.