Fix and clarify usage of references in Spicy.
References weren't fully consistent in their properties and
implementation, and they weren't documented either. This commit cleans
that up and adds documentation.

Changes:

- Constness 1: For a type `T&`, which internally is `strong_ref<T>`,
  we now consistently make the inner `T` a mutable LHS type. That
  seems most natural and useful from a user perspective. In
  particular, this allows mutating objects passed into functions as
  parameters without declaring them `inout`: (`foo (x: bytes&)`). This
  is important because declaring the parameter `inout` makes the
  reference itself mutable (not the contained object), which isn't
  what one wants. Plus, the latter doesn't work for unit parameters
  anyways. Accordingly, we also adapt our built-in operators to
  use`T&` instead of `inout T&`.

- Constness 2: For a type `T&`, we now make the outer reference type
  constant and non-mutable. This is mostly for consistency, it doesn't
  really change anything as we didn't havemutating operations on
  references anyways.

- Feature: We add `new BASIC_TYPE` as syntax for creating default
  initialized values of basic types. So far we only had `new
  NAMED_TYPE` and `new VALUE`.

Technical note to the reviewer: the changes to the `recreateAs*`
methods are necessary to avoid complex types being copied into the
newly created type, which would lead to name resolution problems. This
is tested through existing tests which fail otherwise
(`spicy.types.sink.filter-it.spicy`,
`spicy.rt.base64-filter-eod.spicy`). (Well, `recreateAsLHS()` is
tested that way, but the others would have the same issue.) This is a
good change anyways, because it keeps the AST smaller.

Closes #1618.