Extend stream API to allow for chunks that don't own their data.
By default, we still copy data when creating chunks but we add a
parallel API that just stores pointers, assuming the data will stay
around as long as needed. If the stream owner cannot guarantee that,
they may at any point convert all not-owned data into owned data
through a corresponding `makeOwning()` stream method. To make that
method efficient even with long chains of chunks, we internally
maintain an invariant that only the last chunk of chain can be
non-owning: whenever we add a new chunk to a chain, we ensure that the
previous tail become owning at that point. In other words, we amortize
the work across all newly added chunks.

We'll later use this new API inside the Spicy runtime driver.