Mark automatic derefs inside the operator's AST node.
This allows to differentiate between explicit `deref` operations part
of the source code and implicit `deref` operations inserted by the
coercer. A subsequent commit will leverage this information, but I've
been meaning to do this anyways because it could be hard to track
where a particular `deref` was coming from.