Fix `new` passing a unit reference to an `inout` unit parameter.
One might debate whether this is something we should allow at all but
we do permit it elsewhere when passing unit parameters, and changing
that would probably break code, so for consistency this allows it for
`new` as well.

Internally, there was a more general inconsistency introduced by
automatic derefs of Spicy-level strong references. While we do need
that deref to happen for operator resolution, we now remove it after
generation of HILTI code so that HILTI-level resolution can then work
as expected on the value references introduced for units. I wouldn't
be surprised if the prior behavior was causing more trouble than just
#1719 and we just hadn't run into yet it.

Closes #1719.