Reject joint usage of filters and look-ahead.
We cannot retrieve look-ahead information from units to which a filter
is connected, because parsing for outer and inner layers will be
operating on separate streams. While we could pass the lahead token
number upwards, we would also need the stream position for the end of
the lahead symbol, but we cannot tie an iterator on the filtered
stream back to a position on the original stream. So we now reject
this.

This shouldn't cause any actual backwards-incompatibility because this
wasn't working in the first place: it  would reliably abort with an
invalid stream iterator exception during parsing.

Closes #1655.