Fix skipping in size-contrained units.
We previously could skip too much data if `skip` was used in a unit with
a global `&size`. This was due to the machinery moving the input forward
believing that `skip` units were that really field productions. While
this is true in a sense in this particular case it still lead to
incorrect behavior, in particular as far is input handling is concerned
`skip` productions largely behave like any other unit.

With this patch we store information about field in `Skip` productions.
We can use that additional information to in this case correctly handle
productions for `skip` as well.

Closes #1724.