Fix `&convert` typing issue with bit ranges.
Turns out #1664 was only indirectly related to the `&convert` itself;
the real issue was that we couldn't assign one bitfield struct to
another if their field types didn't match exactly, even in cases where
at the C++ level there was no meaningful difference. In this case we
ended up with a field that had a C++ type `rt::Bool` in one type and
`bool` in another, leading to errors when assigning the latter to the
former. We now allow to creating instances of the former from the
latter through standard C++ type conversions on a per field basis.

Closes #1664.