Use a checked cast for `map`'s `in` operator.
We would previously assume that operand types in `map`'s `in` operator
were always resolved which is not the case, e.g., for this code `x.b`
might be unresolved:

```
public type X = unit {
    var a: map<bytes, bytes>;
    : Y(self);
};

type Y = unit(x: X) {
    on %done { b"" in x.a; }
};
```

The implementation of `set`'s `in` operator already accommodated this by
defering until all types are resolved. This changes the implementation
for `map`'s `in` to follow a similar pattern.

Closes #1743.