Only allow creation of `const` variables from literals.
Since it is hard to enforce that `const` variables are initialized in
correct order (order of decls usually is not significant in Spicy/HILTI
and works e.g., for `global`s, but this is not the case for `const`s due
to codegen), this patch disallows creating `const` values from anything
but literals. This completely removes the ordering issue.

Closes #1763.