Add unit method `stream()` to access current input stream.
We implement this by storing the current input stream inside the
internal parse struct as a new field `__stream`. Using `stream()` then
turns into an access to that field (i.e,  `*(self.__stream)`). If that
field never gets accessed, the optimizer will remove it, so there's no
overhead if nobody ever uses `stream()` with a unit.