Add `%sync_advance` hook.
This adds support for a new unit hook:

```
on %sync_advance(offset: uint64) {
    ...
}
```

This hook is called regularly (see below) during error recovery when
synchronization skips over data or gaps while searching for a valid
synchronization point. It can be used to check in on the
synchronization to, e.g., abort further processing if it just keeps
failing. `offset` is the current position inside the input stream that
synchronization just skipped to.

By default, "called regularly" means that it's called every 4KB of
input skipped over while searching for a synchronization point. That
value can be changed by setting a unit property
`%sync-advance-block-size = <number of bytes>`.

As an additional minor tweak, this also changes the name of what used
to be the `__gap__` profiler to now be called `__sync_advance` because
it's profiling the time spent in skipping data, not just gaps.

Implementation notes:

1. I considered reusing `on %gap()` instead of adding a new hook, but
   it turns out the semantics are different than what I think we want
   here: (1) it doesn't trigger for non-gap data (by definition), so
   couldn't check in on long intervals of searching through payload
   data; and (2) it's called when a gap added to the input, not when
   it's encountered.

2. The new `ASTInfo` inside the Spicy code generator might seem
   overkill for just the one field it currently contains. However, I'm
   planing to use this for other information in the future as well
   (specifically, tracking the need for `&on-heap`).

Closes #3779.