Prevent defining already forwarded constants.
We previously would add both forward declarations as well as full
definitions for any constant in an imported module, i.e. we emited
the same constant in multiple C++ files which the linker seems to have
cleaned up for us.

With this path we only emit a definition if we are not already emitting
a forward declaration.

Closes #1785.