Tweak when we emit struct inline code.
We would previously emit inline C++ code for types into each unit module
where the type's module was imported. This meant that we potentially
compiled identical function bodies in many files. While this could in
theory lead to better local inlining decisions it also caused poor
scaling as the number of modules in a Spicy grammar grows (in
particular).

With this patch we now emit these bodies at most once, in particular
into the module of the type. To support this we make a number of
declarations and definitions not `inline` anymore so they can link
correctly. We also reduce which code gets included when importing a
module.