Guard access to sink hooks.
If no sink hooks were specified by the user trying to invoke them would
raise a `std::bad_function_call` previously since we did not have
explicitly guard the code. With #1736 we us infrastructure with less
implicit checks so we need to handle the absence of a value ourself
since otherwise we might end up dereferencing a nullptr.