Fix non-converging optimizer pass for used functions.
The optimizer's `FunctionVisitor` computes whether a function is used or
required by any feature. Since recently we now hold all modules in a
single AST. This means that we need at least one full pass through the
AST to collect feature constants so we can decided whether a function is
required by a feature and cannot be elided. The initial implementation
of this used explicit control flow, but this is error-prone like seen in

This patch removes the explicit control flow and instead checks whether
we have discovered any new feature constants. With that the optimizer
pass converges again.

Closes #1808.