Disallow expanding limited `View`s again with `limit`.
The documented semantics of `View::limit` are that it creates a new view
with equal or smaller size. In contrast to that we would still have
allowed to expand views with more calls `limit` again as well.

This patch changes the implementation of `View::limit` so it can only
ever make a `View` smaller.

We also tweak the implementation of the check for consumed `&size` when
used together with `&eod`: if the `&size` was already nested in a
limited view a larger `&size` value could previously extend the view so
the `&eod` effectively was ignored. Since we now do not extend the
`View` anymore we need to only activate the check for consumed `&size`
if `&eod` was not specified since in this case the user communicated that
they are fine with consuming less data.

Closes #1815.

(cherry picked from commit 44d87b21c33c63a2c2e39f9ffe22822fdee55fd0)