Add `string::split` and `string::split1` helpers
This doesn't do any work to fix the TODOs that were already in the Bytes
version. Also, there is a small unclear bit about the documentation for
the `split` op in both Bytes and now string, namely that an empty `sep`
argument should make the break take place at sequences of whitespace,
but it does not do that. It may also be misreading the documentation,
but one of the two should probably be clarified.