Fix `split` and `split1` docs to be more accurate
The documentation for the bytes method seemed incorrect (or, at least,
confusing) where an empty separator did *not* split on white space. This
correctly documents the behavior with what the actual behavior is
(especially for split1).

Note, the behavior is a bit confusing, unintuitive, and maybe
inconsistent between the implementations of `split` and `split1`.
Namely, `split1` will sometimes but the entire value in the first
element of the tuple (if a match isn't found) or the second element (if
the separator is just an empty string).

For now this just updates the documentation, but I'd also welcome a
change in behavior.