Avoid use-after-free warning with GCC-13.2.
Not exactly sure what's going on here. I *think* calling data->begin()
and data->end(), it sees/assumes one path freeing the underlying chain
of the SafeConstIterator and the other then accessing the freed pointer.
Maybe something about the evaluation of the arguments being unspecified.
It doesn't trigger when fetching offsets sequentially beforehand, so do
that...

    inlined from 'spicy::rt::detail::printParserState(...)::<lambda()>'
      at /src/spicy/runtime /src/parser.cc:81:19:
    /src/hilti/runtime/include/hilti/rt/intrusive-ptr.h:43:20:
      error: pointer used after 'void operator delete(void*, std::size_t)'
      -Werror=use-after-free]
    43 |     if ( m && --m->_references == 0 )
       |                 ~~~^~~~~~~~~~~$